mvp/requirements.txt
fastapi
sqlalchemy
uvicorn
psycopg2-binary
python-multipart
requests
streamlit
streamlit_option_menu
transformers
torch
BeautifulSoup4